Added contracts, working on add() implementation
Added contracts to track progress
Thanks to Atticus K for the block comment contract schemes.
Add is proving difficult to implement. I am considering using a helper
function and keeping track of the parent node but I am confused as to
how I should return an entire tree with the new node added…